Licenta2017LuncasuBogdan-Cristian/objrec/requirements.txt
Flask<1
beautifulsoup4==4.5.3
Flask-Cors==3.0.2
MarkupSafe==0.23
matplotlib==2.0.0
numpy==1.11.3
opencv-python==3.2.0.6
Pillow==4.0.0
requests==2.13.0
scipy==0.18.1
SQLAlchemy==1.1.9
Werkzeug==0.11.15
pathlib==1.0.1
psycopg2>=2.7